+++ FuzzyString/OverlapCoefficient.cs
@@ -5,9 +5,9 @@ namespace FuzzyString
 {
     public static partial class ComparisonMetrics
     {
-        public static double OverlapCoefficient(this string source, string target)
+        public static float OverlapCoefficient(this string source, string target)
         {
-            return (Convert.ToDouble(source.Intersect(target).Count())) / Convert.ToDouble(Math.Min(source.Length, target.Length));
+            return Convert.ToSingle(source.Intersect(target).Count()) / Convert.ToSingle(Math.Min(source.Length, target.Length));
         }
     }
 }
