+++ FuzzyString/RatcliffObershelpSimilarity.cs
@@ -5,9 +5,9 @@ namespace FuzzyString
 {
     public static partial class ComparisonMetrics
     {
-        public static double RatcliffObershelpSimilarity(this string source, string target)
+        public static float RatcliffObershelpSimilarity(this string source, string target)
         {
-            return (2 * Convert.ToDouble(source.Intersect(target).Count())) / (Convert.ToDouble(source.Length + target.Length));
+            return 2 * Convert.ToSingle(source.Intersect(target).Count()) / Convert.ToSingle(source.Length + target.Length);
         }
     }
 }
