+++ FuzzyString/TanimotoCoefficient.cs
@@ -4,11 +4,11 @@ namespace FuzzyString
 {
     public static partial class ComparisonMetrics
     {
-        public static double TanimotoCoefficient(this string source, string target)
+        public static float TanimotoCoefficient(this string source, string target)
         {
-            double Na = source.Length;
-            double Nb = target.Length;
-            double Nc = source.Intersect(target).Count();
+            float Na = source.Length;
+            float Nb = target.Length;
+            float Nc = source.Intersect(target).Count();
 
             return Nc / (Na + Nb - Nc);
         }
